+++ MinkowskiEngine/MinkowskiCoordinateManager.py
@@ -238,6 +238,9 @@ class CoordinateManager:
     ):
         return self._manager.exists_field_to_sparse(field_map_key, sparse_map_key)
 
+    def field_to_sparse_keys(self, field_map_key: CoordinateMapKey):
+        return self._manager.field_to_sparse_keys(field_map_key.get_key())
+
     def get_field_to_sparse_map(
         self, field_map_key: CoordinateMapKey, sparse_map_key: CoordinateMapKey
     ):